Added CWT

+++ 01-recipes/index.js
@@ -16,6 +16,38 @@ function generateAccessToken(id) {
         expiresIn: "3w"
     });
 }
+// a middelware function to verify the JWT token, it happens before the route handler is called
+// the next parameter will refer to the next middleware
+// or if there is no more middleware, thhe route handler will be called
+function verifyToken(req, res, next) {
+
+    // the token is sent in the Authorization header, in the format of "Bearer <token>"
+    const authHeader = req.headers['authorization'];
+    if (authHeader) {
+        const token = authHeader.split(' ')[1];
+
+        if (token) {
+            // verify the token's claims and expiry matches the signiture
+            jwt.verify(token, process.env.TOKEN_SECRET, function(err,claims) {
+                if (err) {
+                    res.status(403).json({
+                        "message": "Invalid token"
+                    });
+                } else {
+                    // save in the request the logged in user, so that the route handler can use it
+                    req.user = claims;
+                    next();
+                }
+            });
+        }
+
+    } else {
+        res.status(400).json({
+            "message": "Authorization header not found"
+        })
+    }
+    
+}
 
 const app = express();
 
@@ -70,7 +102,7 @@ async function main() {
 
     })
 
-    app.post("/api/recipes", async function(req, res){
+    app.post("/api/recipes", [verifyToken], async function(req, res){
         // add the new recipe from the request's body
         const newRecipe = req.body;
         
@@ -85,6 +117,8 @@ async function main() {
         // replace the newRecipe's cuisine and tags with the ones from the database
         newRecipe.cuisine = cuisine;
         newRecipe.tags = tags;
+
+        newRecipe.user_id = new ObjectId(req.user.user_id); // associate the recipe with the logged in user
         
         const response = await db.collection("recipes").insertOne(newRecipe);
         res.json({
@@ -93,30 +127,43 @@ async function main() {
         });
     })
 
-    app.delete("/api/recipes/:recipeId", async function(req, res){
+    app.delete("/api/recipes/:recipeId", [verifyToken], async function(req, res){
         const recipeId = req.params.recipeId;
-
+    
+        // find the recipe by its id and the user_id of the logged in user, to make sure the user can only delete their own recipes
+        const recipe = await db.collection("recipes").findOne({
+            user_id: new ObjectId(req.user.user_id),
+            _id: new ObjectId(req.params.recipeId)
+        });
+        if (recipe) {
         // delete one document which has the _id equal to the recipeId
         const result = await db.collection("recipes").deleteOne({
              _id: new ObjectId(recipeId) 
             })
         res.json({
             message: "Recipe deleted successfully"
-        });
+        })
+        } else {
+        res.status(400).json({
+            message: "Recipe not found, or you do not owe the recipe"})
+        }
     })
 
     // PATCH or PUT
-    app.put("/api/recipes/:recipeId", async function(req, res){
+    app.put("/api/recipes/:recipeId", [verifyToken], async function(req, res){
         // get the new recipe from the request's body
         const newRecipe = req.body;
-       
+        
+        const recipe = await db.collection("recipes").findOne({
+            user_id: new ObjectId(req.user.user_id),
+            _id: new ObjectId(req.params.recipeId)
+        })
         // find the cuisine and associate the recipe with the cuisine
         const cuisine = await db.collection("cuisines").findOne({ 
             name: newRecipe.cuisine });
 
         const tags = await db.collection("tags").find({ 
-            name: { 
-                $in: req.body.tags } 
+            name: { $in: req.body.tags } 
         }).toArray();
         
         // replace the newRecipe's cuisine and tags with the ones from the database
@@ -130,9 +177,9 @@ async function main() {
         });
         res.json({
             message: "Successfully updated the recipe",
-        });
+        })
 
-    });
+    })
 
     // register a new user
     // shape of req.body: { username: string, password: string }
@@ -163,19 +210,30 @@ async function main() {
                     res.json({
                         "message": "Login successful",
                         "token": token
-                    });
+                    })
                 } else {
                     res.status(401).json({
                         message: "Wrong email or password"
-                    });
+                    })
                 }
             } else {
                 res.status(401).json({
                     message: "Wrong email or password"
-                });
+                })
             }
         })
-    }
+    app.get("/api/me", verifyToken, async function(req, res){
+        // the req.user is set by the verifyToken middleware
+        const user = await db.collection("users").findOne({ 
+            _id: new ObjectId(req.user.user_id) 
+        });
+        delete user.password; // remove the password from the user object before sending it back
+        res.json({
+                "user": user
+            })
+    })
+
+}
 
 main();
